Fiexd vm and request models, added check NullReferebceException

+++ MachineWarehouse/Services/CarServices/CarServices.cs
@@ -18,16 +18,9 @@ namespace MachineWarehouse.Services.CarServices
             _mapper = mapper;
         }
 
-        public async Task<Car> AddCar(CreateCar request)
+        public async Task<Car> AddCar(CarRequests request)
         {
-            var car = new Car
-            {
-                YearRelese = request.YearRelese,
-                Price = request.Price,
-                ShorDescription = request.ShorDescription,
-                BrandId = _context.Brands.FirstOrDefault(brand => brand.Brand == request.BrandName).Id,
-                ColorId = _context.Colors.FirstOrDefault(color => color.Color == request.ColorName).Id,
-            };
+            var car = _mapper.Map<Car>(request);
 
             await _context.Cars.AddAsync(car);
             await _context.SaveChangesAsync();
@@ -39,33 +32,50 @@ namespace MachineWarehouse.Services.CarServices
         {
             var car = await _context.Cars.FindAsync(id);
 
+            if(car == null) 
+            {
+                throw new Exception("Автомобиль не найден");
+            }
+
             _context.Cars.Remove(car);
             await _context.SaveChangesAsync();
         }
 
-        public async Task<List<GetCarsVm>> GetAllCars()
+        public async Task<List<CarVm>> GetAllCars()
         {
-            var cars =  _mapper.Map<List<GetCarsVm>>(await _context.Cars.ToListAsync());
+            var cars =  _mapper.Map<List<CarVm>>(await _context.Cars.ToListAsync());
 
             return cars;
         }
 
-        public async Task<GetCarVm> GetCarById(int id)
+        public async Task<CarVm> GetCarById(int id)
         {
-            var car = _mapper.Map<GetCarVm>(await _context.Cars.FirstOrDefaultAsync(car => car.Id == id));
+            var entity = await _context.Cars.FirstOrDefaultAsync(car => car.Id == id);
+
+            if (entity == null) 
+            {
+                throw new Exception("Автомобиль не найден");
+            }
+
+            var car = _mapper.Map<CarVm>(entity);
 
             return car;
         }
 
-        public async Task UpdateCar(UpdateCar request)
+        public async Task UpdateCar(CarRequests request)
         {
             var entity = await _context.Cars.FirstOrDefaultAsync(car => car.Id == request.Id);
+
+            if (entity == null) 
+            {
+                throw new Exception("Автомобиль не найден");
+            }
 
             entity.YearRelese = request.YearRelese;
             entity.Price = request.Price;
             entity.ShorDescription = request.ShorDescription;
-            entity.BrandId = _context.Brands.FirstOrDefault(brand => brand.Brand == request.BrandName).Id;
-            entity.ColorId = _context.Colors.FirstOrDefault(color => color.Color == request.ColorName).Id;
+            entity.BrandId = request.BrandId;
+            entity.ColorId = request.ColorId;
 
             await _context.SaveChangesAsync();
         }
